Fixed a bug in the bookkeeping.
When booking a number, GFBookkeeping::BookNumbers() wrongly tested fPlanes
instead of fNumbers if the key was already there and thereby created a new key
in the map with an arbitrary pointer. This led to a segmentation fault in the
copy constructor, when it tries to copy all the planes for this wrong key.

git-svn-id: svn+ssh://svn.code.sf.net/p/genfit/code/trunk@450 4ab16672-d4cc-4d65-938a-feac61c82957