Introduced GFTools::updateRepSmoothed().
This function allows to update a track representation with the data stored by
smoothing. It is provided with the track, the number of the representation and
the number of the hit and then updates the reference plane, the state vector and
the covariance matrix of the representation to the data that was saved for the
given hit during the fit.


git-svn-id: svn+ssh://svn.code.sf.net/p/genfit/code/trunk@505 4ab16672-d4cc-4d65-938a-feac61c82957